Use correct LICENSE file.
Our intention to license our changes under the ALv2 while
respecting the original work under a 3 clause BSD-esque license.
This is a totally normal thing to do with permissively licensed code
but the exact mechanics are not agreed upon. The closest thing to a
pattern is "put your license on top, and everything else somewhere
below". The original authors are noted in LICENSE, NOTICE, and README
files.